added image background generation.
