+++ templates/ios/XcodeProjectTemplate/projectName/DesignToCode/Configs/containerNameConfig.swift.hbs
@@ -78,8 +78,8 @@ class {{container.name}}Config: ViewConfigImpl {
         {{/if}}
         {{/eq}}
         {{#eq type 'Image'}}
-        view = UIImageView(frame: CGRect(x: {{rect.x}}, y: {{rect.y}}, width: {{rect.width}}, height: {{rect.height}}))
-        (view as! UIImageView).image = UIImage(named: "DtcGenerated/images/{{stem imageName}}")
+        view = ImageView(frame: CGRect(x: {{rect.x}}, y: {{rect.y}}, width: {{rect.width}}, height: {{rect.height}}))
+        (view as! ImageView).image = UIImage(named: "DtcGenerated/images/{{stem imageName}}")
         view!.contentMode = .scaleAspectFill
         {{#if backgroundColor }}
         view!.backgroundColor = UIColor(red: {{backgroundColor.fill.red}}, green: {{backgroundColor.fill.green}}, blue: {{backgroundColor.fill.blue}}, alpha: {{backgroundColor.fill.alpha}})
